refactor: ticket single: reuse `TopBar.vue` (#1300)
